fix: consolidate EmptyState to single component (#1174)
- Remove duplicate components/EmptyState.tsx (FontAwesome version)
- Keep components/ui/EmptyState.tsx as the single source of truth
- Make icon prop optional with default "inbox" for backward compat
- Update 10 screen imports from @/components/EmptyState to @/components/ui/EmptyState
